Move blueprints.types -> blueprints.variables.types
